requirements.txt
streamlit>=1.36.0
google-genai>=0.3.0
google-generativeai>=0.7.2
